Drop 74ABT74
The 74ABT74 is only used as a divider to create 25MHz from the 50MHz clock, and
it can be removed. Instead, the 50MHz are directly brought to the ATF1508. CLK
will be used as an output, not an input, so move away from GCLK pin. CLKI is no
longer needed on a pin as it doesn't connect to anything anymore.